backend/requirements.txt
# requirements.txt - Python Dependencies for Nexxa Backend

# Django Core
Django==5.0
djangorestframework==3.14.0

# Database
mysqlclient==2.2.0

# CORS Headers
django-cors-headers==4.3.1

# JWT Authentication
PyJWT==2.8.0
cryptography==41.0.7

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Optional but recommended for production
gunicorn==21.2.0
whitenoise==6.6.0
django-environ==0.11.2

# Development tools (optional)
ipython==8.18.0
django-debug-toolbar==4.2.0
